Add ALLOW_NESTED_REPEAT syntax option
Not all languages allow nested repeats (/a***/, /a{1}{1,}/, etc) in
their regular expression handling. This adds an additional Syntax
behavior to control whether nested repeats should be allowed. The
default behavior, GNU_REGEX_BV, was updated to allow nested repeats.

This also required modifying the logic that looks at the return value
of QuantifierNode.setQuantifier to update the repeat's target even
when the repeat option given was "{1}" or "{1,1}".