Add a lightweight region for numRegs == 1
Whie benchmarking the CSV library in JRuby, one of the largest
groups of allocated objects was int[] primarily for Region
objects, some of which most likely only have one regionand do not
need the flexibility of an array of regions. This commit splits
Region into multi and single forms, so that single-region objects
can be as compact as possible and not require a second (or third)
object for the beg/end arrays.