Move to a factory method for Region instances
In order to specialize Region based on width, we need consumers to
stop instantiating it directly. This commit deprecates the two
Region constructors, replacing them with newRegion factory methods
we can use in the future to construct specialized Region shapes.

This feature will have to be in the wild for some time to allow
JRuby and libraries like strscan to move to the updated API. At
some point in the future, we can major rev joni and remove the
deprecated constructors and public fields.