Provide encoding for group name exception
Without an encoding provided, this code assumes the bytes are in
the default character set. When they are not, the decoded string
will be corrupt. This is the cause of a failure in the strscan
library on Windows.

This commit adds encoding to the parameters for named group lookup
so that it can be decoded properly if an exception must be raised.