Pool a matcher per thread
Matching is a leaf operation, so we can pool a matcher per thread
to avoid the cost of constructing and initializing the entire
matcher for each string.

This patch does eliminate all object creation in typical usage,
reusing all objects and expecting users to pull match data out of
them before using the regex again. However in benchmarks of a
regex-heavy StringScanner-based CSV library (jruby/jruby#7604) the
overhead of fetching and clearing the pooled matcher appears to
hurt performance more than it helps to remove the allocations.